break module into two parts
