+++ src/isacalc/livingFund.js
@@ -1,0 +1,13 @@
+import {
+  isaSessionEndDate,
+  isaSessionStartDate,
+  momentDayOnly,
+  numDaysInISASession,
+  openDaysBetween,
+} from '@learnersguild/guild-dates'
+
+// --- living fund
+
+export const LIVING_FUND_STIPEND_AMOUNT = 13846
+export const LIVING_FUND_ISA_MAX_PERCENTAGE = 0.085 // 8.5%
+export const LAPTOP_STIPEND_AMOUNT = 1846
